misc cleanup/organization
moved common-ish functions to utils or common. Removed unnecessary code.
